+++ src/runtime/pynumber.cs
@@ -5,7 +5,6 @@ namespace Python.Runtime
     /// <summary>
     /// Represents a generic Python number. The methods of this class are
     /// equivalent to the Python "abstract number API". See
-    /// PY2: https://docs.python.org/2/c-api/number.html
     /// PY3: https://docs.python.org/3/c-api/number.html
     /// for details.
     /// </summary>
@@ -17,11 +16,6 @@ namespace Python.Runtime
         protected PyNumber(IntPtr ptr) : base(ptr)
         {
         }
-
-        protected PyNumber()
-        {
-        }
-
 
         /// <summary>
         /// IsNumberType Method